refactor: Changed lexer transform names
