+++ src/txxt/lexer/transformations.rs
@@ -15,17 +15,17 @@
 //! - experimental_pipeline - orchestrates all transformations into cohesive pipeline
 
 pub mod experimental_pipeline;
-pub mod experimental_transform_indentation_to_token_tree;
-pub mod experimental_transform_to_line_tokens;
 pub mod transform_blanklines;
 pub mod transform_indentation;
+pub mod transform_indentation_to_token_tree;
+pub mod transform_to_line_tokens;
 pub mod transform_whitespace;
 
 pub use experimental_pipeline::{
     experimental_lex, experimental_lex_stage, PipelineOutput, PipelineStage,
 };
-pub use experimental_transform_indentation_to_token_tree::experimental_transform_indentation_to_token_tree;
-pub use experimental_transform_to_line_tokens::experimental_transform_to_line_tokens;
 pub use transform_blanklines::transform_blank_lines;
 pub use transform_indentation::transform_indentation;
+pub use transform_indentation_to_token_tree::experimental_transform_indentation_to_token_tree;
+pub use transform_to_line_tokens::experimental_transform_to_line_tokens;
 pub use transform_whitespace::process_whitespace_remainders;
